requirements.txt
abstract==2021.11.29.1
asgiref==3.5.0
backports.zoneinfo==0.2.1
base32hex==1.0.2
certifi==2021.10.8
charset-normalizer==2.0.12
colouration==2020.7.26
company==0.0.4
Django==4.0.3
django-cors-headers==3.12.0
django-environ==0.9.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
graphviz==0.19.1
idna==3.3
pickle-warehouse==0.1.1
PyJWT==2.4.0
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
